simulation/requirements.txt
# Simulation Engine
ray[default]==2.9.0
paho-mqtt==1.6.1
chromadb==0.4.22
google-generativeai==0.3.2
tenacity==8.2.3
numpy==1.26.3
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
